requirements.txt
jinja2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart
httpx==0.25.2
sqlalchemy==2.0.23
python-telegram-bot==20.7
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0
requests==2.31.0
aiofiles==23.2.1
